pyproject.toml
[project]
name = "schemax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.129.0",
    "passlib>=1.7.4",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
